+++ frontend/src/metabase/modes/components/drill/CompareToRestDrill.js
@@ -1,30 +1,14 @@
 import { t } from "ttag";
 
 import MetabaseSettings from "metabase/lib/settings";
-import { isExpressionField } from "metabase-lib/lib/queries/utils/field-ref";
+import {
+  compareToRestDrill,
+  compareToRestDrillUrl,
+} from "metabase-lib/lib/queries/drills/compare-to-rest-drill";
 
 export default ({ question, clicked }) => {
-  const query = question.query();
-  if (!question.isStructured() || !query.isEditable()) {
-    return [];
-  }
-
-  // questions with a breakout
-  const dimensions = (clicked && clicked.dimensions) || [];
-
-  // ExpressionDimensions don't work right now (see metabase#16680)
-  const includesExpressionDimensions = dimensions.some(dimension => {
-    return isExpressionField(dimension.column.field_ref);
-  });
-
-  const isUnsupportedDrill =
-    !clicked ||
-    dimensions.length === 0 ||
-    // xrays must be enabled for this to work
-    !MetabaseSettings.get("enable-xrays") ||
-    includesExpressionDimensions;
-
-  if (isUnsupportedDrill) {
+  const enableXrays = MetabaseSettings.get("enable-xrays");
+  if (!compareToRestDrill({ question, clicked, enableXrays })) {
     return [];
   }
 
@@ -35,15 +19,7 @@ export default ({ question, clicked }) => {
       icon: "bolt",
       buttonType: "token",
       title: t`Compare to the rest`,
-      url: () => {
-        const filters = query
-          .clearFilters() // clear existing filters so we don't duplicate them
-          .question()
-          .drillUnderlyingRecords(dimensions)
-          .query()
-          .filters();
-        return question.getComparisonDashboardUrl(filters);
-      },
+      url: () => compareToRestDrillUrl({ question, clicked }),
     },
   ];
 };
